Starting to add i18n to core.
--HG--
branch : 0.7
rename : tests/test_ext_i18n.py => tests/test_i18n.py
rename : tipfyext/i18n.py => tipfy/i18n.py
